plugins: Fix libmicrohttpd create response memory mode used
When calling MHD_create_response_from_buffer() function to create
a response for an HTTP GET request, the memory mode type used was
MHD_RESPMEM_PERSISTENT.

This is clearly wrong since the buffer with the data to send back
to the client is allocated from the requestCallback() stack.

Instead MHD_RESPMEM_MUST_COPY should be to let libmicrohttpd copy
the buffer so it isn't freed before the response can be sent.