plugins: Make commands to move RoDI ad infinitum
RoDI has an unusual behaviour in the sense that a move command makes the
robot to keep moving until a new move command is received. This decision
was made to simplify it's programming model for the kids using the robot.

The Gazebo Joint class SetVelocity() function on the other hand just sets
the velocity for a single iteration in the simulation. So the move values
need to be cached and the velocity set for both joints on each iteration.