plugins: Use proper values for the wheels joints angular velocity
The move command sends a value on a range between -100 and 100. These are
values that indicate a percentage of the servos maximum operating angle.

So -100..100 maps to -180..180 degrees. But the Joint class SetVelocity()
function takes radians per second as argument.

Instead of using invented values, calculate the proper rad/s values and
update the joints properly.