models: Add mass, inertia vector and friction to RoDI links
The model didn't have a mass, inertia vector and friction values for the
base, ball caster and left/right wheels. This made the robot to slip and
didn't stop after a stop command.

RoDI has a weight of 115 grams and following the same mass distribution
for similar robots, 33% of its mass is on each wheel, 16% in the ball
caster and 18% in the base. The inertial vectors values were calculated
according to the links mass and link geometry.