plugins: Fix termination due wrong exception throw in RodiWeb constructor
If a libmicrohttpd webserver fails to be started, the RodiWeb() constructor
throws an exception, but just uses the throw keyword without any expression.

When using the throw keyword without an expression, the C++ runtime attempts
to rethrow the currently handled exception. But since MHD_start_daemon() is
a C function and of course doesn't throw an exception, the C++ runtime calls
std::terminate() since exception handling fails.

This causes the Gazebo simulator to terminate with the following error:

"terminate called without an active exception"

Fix this by using the gzthrow() macro that not only throws an exception with
useful debug information such as the plugin and function where happened, but
also prints this info and the one provided to the macro in the throw stream.

Unfortunately the MHD_start_daemon() function just returns NULL in case of
an error, so the exact error cause can't be added to the Gazebo exception.