models: Decouple the RoDI plugin from the model
Including the RoDI plugin in the model is convenient since only the
latter has to be included in a world file.

But unfortunately there is no way in Gazebo to override XML tags in
an included model, so having the plugin in the model doesn't allow
to change the TCP port for the HTTP server.

This is needed not only for customization purposes but also to have
more than one RoDI model in a simulated world.