plugins: Fix sonar name used for sensor lookup
Commit b76f81143c86 ("models: Decouple the RoDI plugin from the model")
removed the RoDI plugin from the model so multiple robots could be used
in a simulated world.

But when the plugin isn't in the model the elements are namespaced with
the model name. Only the left and right wheel joints names were updated
in the mentioned commit, but not for the sonar which lead to all robots
using the same sonar.

Fix this by first looking up the base_link that contains the sonar and
then get the sensor absolute name path so the SensorManager can get it.