plugins: Sanity check the included model on RoDI plugin Load()
The RoDI plugin assumes the used SDF is correct, but not including the
proper model://rodi will lead to a NULL pointer dereference error.

Add some sanity check verifying the used links at least are present in
the included model and don't start the web server if is not the case.