The light sensor and the IR sensor have been added. The buzzer has been implemented but commented because it is not cross-platform. The light sensor uses a camera to take a picture, it calculates the average color of the image and uses the formula Luminance = 0.2126*Red + 0.7152*Green + 0.0722*Blue. Then the console outputs the value which has a range 0-1023 (0 being the darkest and 1023 the brightest) The IR sensor does the same as the light sensor, but the sensor points and simulates a sensor that calculates the reflectiveness of the lower surface. The image size is much smaller as well. The buzzer has been implemented but commented out. Future development may include a cross-platform way of doing this.
The light sensor uses a camera to take a picture, it calculates the average color of the image and uses the formula Luminance = 0.2126*Red + 0.7152*Green + 0.0722*Blue. Then the console outputs the value which has a range 0-1023 (0 being the darkest and 1023 the brightest)
The IR sensor does the same as the light sensor, but the sensor points and simulates a sensor that calculates the reflectiveness of the lower surface. The image size is much smaller as well.
The buzzer has been implemented but commented out. Future development may include a cross-platform way of doing this.